fix(desktop): split env.ts to prevent electron import in terminal-host daemon
The terminal-host daemon runs with ELECTRON_RUN_AS_NODE=1 and cannot
resolve the electron module. Vite's code splitting bundled electron-dependent
code (via local-db) into a shared chunk loaded by terminal-host, causing
immediate crash on startup (ELECTRON-X, ELECTRON-Y, ELECTRON-10).

Extract buildTerminalEnv and its electron/local-db dependencies into
env-terminal.ts, keeping env.ts electron-free so terminal-host/session.ts
can safely import buildSafeEnv without pulling in electron.
